research_aggregator/ingest/requirements.txt
numpy<2.0
arxiv
PyYAML
python-dotenv
transformers
keybert
torch
sentence-transformers
supabase